requirements.txt
# Core dependencies
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0

# CLIP dependencies
ftfy>=6.1.0
regex>=2023.0.0
